feat: add keyword check
